refactor(lambda): move reusable functions from handler to helper file
